refacto: split qprofiles-manager in 2 jobs
