+++ qgis_deployment_toolbelt/jobs/job_profiles_synchronizer.py
@@ -1,7 +1,8 @@
 #! python3  # noqa: E265
 
 """
-    Download remote QGIS profiles to local.
+    Synchronize profiles between downloaded (in QDT working folder) and installed
+        profiles (in QGIS profiles'folder).
 
     Author: Julien Moura (https://github.com/guts)
 """
@@ -19,9 +20,7 @@ from shutil import copy2, copytree
 
 # package
 from qgis_deployment_toolbelt.jobs.generic_job import GenericJob
-from qgis_deployment_toolbelt.profiles import LocalGitHandler, RemoteGitHandler
 from qgis_deployment_toolbelt.profiles.qdt_profile import QdtProfile
-from qgis_deployment_toolbelt.profiles.remote_http_handler import HttpHandler
 
 # #############################################################################
 # ########## Globals ###############
@@ -36,48 +35,14 @@ logger = logging.getLogger(__name__)
 # ##################################
 
 
-class JobProfilesDownloader(GenericJob):
+class JobProfilesSynchronizer(GenericJob):
     """
-    Job to download remote profiles and set them.
+    Job to synchronize profiles between downloaded (in QDT working folder) and installed
+        profiles (in QGIS profiles'folder).
     """
 
-    ID: str = "qprofiles-manager"
+    ID: str = "qprofiles-synchronizer"
     OPTIONS_SCHEMA: dict = {
-        "action": {
-            "type": str,
-            "required": False,
-            "default": "download",
-            "possible_values": ("download", "refresh"),
-            "condition": "in",
-        },
-        "branch": {
-            "type": str,
-            "required": False,
-            "default": "master",
-            "possible_values": None,
-            "condition": None,
-        },
-        "local_destination": {
-            "type": str,
-            "required": False,
-            "default": ".cache/qgis-deployment-toolbelt/profiles",
-            "possible_values": None,
-            "condition": None,
-        },
-        "protocol": {
-            "type": str,
-            "required": True,
-            "default": "git_remote",
-            "possible_values": ("git", "git_local", "git_remote", "http"),
-            "condition": "in",
-        },
-        "source": {
-            "type": str,
-            "required": True,
-            "default": None,
-            "possible_values": ("https://", "http://", "git://", "file://"),
-            "condition": "startswith",
-        },
         "sync_mode": {
             "type": str,
             "required": False,
@@ -110,60 +75,6 @@ class JobProfilesDownloader(GenericJob):
 
     def run(self) -> None:
         """Execute job logic."""
-        # download or refresh
-        if self.options.get("action", "download") != "download":
-            raise NotImplementedError
-
-        # prepare remote source
-        if self.options.get("protocol") in ("git", "git_local", "git_remote"):
-            if self.options.get("protocol") == "git":
-                logger.warning(
-                    DeprecationWarning(
-                        "'git' protocol has been split into 2 more explicit: 'git_local' "
-                        "(for git repositories accessible directly through file system "
-                        "and 'git_remote' (for repositories accessible network request "
-                        "to a remote git server through HTTP). "
-                        "Please update your scenario consequently."
-                    )
-                )
-            if self.options.get("protocol") == "git_remote" or self.options.get(
-                "source"
-            ).startswith(("git://", "http://", "https://")):
-                downloader = RemoteGitHandler(
-                    source_repository_url=self.options.get("source"),
-                    branch_to_use=self.options.get("branch", "master"),
-                )
-            elif self.options.get("source").startswith("file://"):
-                downloader = LocalGitHandler(
-                    source_repository_path_or_uri=self.options.get("source"),
-                    branch_to_use=self.options.get("branch", "master"),
-                )
-            else:
-                logger.error(
-                    f"Source type is not implemented yet: {self.options.get('source')}"
-                    f"for '{self.options.get('protocol')}' protocol"
-                )
-                raise NotImplementedError
-        elif self.options.get("protocol") == "http":
-            if not self.options.get("source").startswith(("http://", "https://")):
-                logger.error(
-                    f"Source type not implemented yet: {self.options.get('source')} "
-                    f"for '{self.options.get('protocol')}' protocol"
-                )
-                raise NotImplementedError
-            downloader = HttpHandler(
-                source_repository_path_or_uri=self.options.get("source"),
-            )
-        else:
-            logger.critical(
-                f"Protocol '{self.options.get('protocol')}' is not part of supported ones: "
-                f"{self.OPTIONS_SCHEMA.get('protocol').get('possible_values')}"
-            )
-            raise NotImplementedError
-
-        # run download operation
-        downloader.download(destination_local_path=self.qdt_downloaded_repositories)
-
         # check of there are some profiles folders within the downloaded folder
         profiles_folders = self.list_downloaded_profiles()
         if profiles_folders is None:
@@ -369,12 +280,3 @@ class JobProfilesDownloader(GenericJob):
                 copy_function=copy2,
                 dirs_exist_ok=True,
             )
-
-
-# #############################################################################
-# ##### Stand alone program ########
-# ##################################
-
-if __name__ == "__main__":
-    """Standalone execution."""
-    pass
